Changed variables from float to double

+++ week2/main.cpp
@@ -11,7 +11,7 @@ using namespace std;
 int main()
 {
     // Declare variables for each of the inputs and the output
-    float input1, input2, input3, input4, input5, outputMean;
+    double input1, input2, input3, input4, input5, outputMean;
 
     //Prompt the user for each of the necessary inputs
     cout << "Please enter five numbers." << endl;
